Wip of the combo logic.
Most of the core logic is implemented.
Some edge cases aren't, though, for instance expanding the second
phase of the combo over the limit doesn't count as a failure yet,
the additionnal takes are simply counted as the first phase
of a new combo.
The difficulty is modulated by flooring the current mult divided by ten.
That mean the difficulty drop to 0 when the user fail during a combo.
Spawning pattern are picked randomly with a seeded random.
The seed is defined in the constant.